Change versioning in master branch builds
It causes some confusion when demos built from the master branch show,
for example, "v2.3.0-uncompiled", even though v2.3.x is in a separate
branch and we are working toward v2.4.0.

Now master branch builds are tagged with the most recent release and
the word "master".  The "master" branch commits corresponding to
branched release versions must now be tagged in addition to the
release itself.

Closes #1420

Change-Id: I4dd79ea7c9a479c7ad9074bde7e6066e4315ef43