Re-add DASH-IF asset.
Previously, we have removed the Big Buck Bunny DASH-IF from our
demo assets. It was not loading properly, due to header issues.
Just today, the DASH-IF people moved those test assets to a new domain
with working headers, so this re-adds the asset with the new domain.

Backported to v2.3.x

Change-Id: I089a781d8525a47dfcadcae2feccacbbdc81f848