Fix bug with merging a period into itself
When we try to merge two SegmentIndex objects, we check to make sure
that they are sequential.  Then, after merging the various SegmentIndexes,
they are fit.
This causes problems when trying to merge two identical periods together,
an operation that could potentially happen in a live stream.
SegmentIndex.merge was comparing the post-fit version of one SegmentIndex
with the pre-fit version of another, causing it to believe that the two
SegmentIndex had different end times when really they were identical.

This change causes SegmentTemplate.createStream to perform an extra fit
operation before merging, to avoid that problem.

Closes #1448

Change-Id: If2e9a9a91de344b7ab8cd9004da0553b324cb599