Silence DRM engine errors for unencrypted assets
We still load up EME for unencrypted assets if configured.
This helps us mix unencrypted ads with encrypted content.

However, this had the side effect that if the user sets DRM
configuration, the player would try to load up EME systems for
unencrypted assets, which would make it fail on platforms with no EME
systems.
That failure shouldn't really be fatal, so this silences the errors in
that case.

Issue #1479

Change-Id: I01a9a3dad8a7f7974e54a6424be8a6515a8e5f99