Isolate Use of Track Cues In Simple Text Displayer
Since we only need to know about which cues a text track has when we
remove cues, there is no need to keep a local copy.

Since some platforms return null if the track is disabled, we do need
to temporarily change the mode if the track is disabled so that we can
get the cues.

However, this limits the scope of these "special" conditions to a single
function.

Change-Id: I7026e9c73629ef86633d5fe732f21a48178e8767