Stop Playhead From Passing Manifest To Anyone
In an effort to decouple Manifest from Playhead this change removes
the utility function getRebufferingGoal.

getRebufferingGoal was only used in playhead.js, playhead_observer.js, and
streaming_engine.js. However, it was used in two different ways:
  - playhead.js and playhead_observer.js always used it with a scale of
    one.
  - streaming_engine.js used it with a custom scale and then with another
    buffering value.

Since the logic inside the function was so simple and the use of the function
between call sites were different enough, there appeared to be little value
in keeping the function.

Change-Id: Ie0983a1f84516a3cf5de34cd5ed87d1e786862d8