Add mime type option to Player.load
Added an option for the mime type to be passed to the load function.
Going forward this will replace the option to pass in a manifest
parser factory.

This is to help with issue #816 as passing in the mime type will be
one way for the player to know if it should use src= over MSE.

Issue #816

Change-Id: Ic356bb9f52c3977620debc8d4a09b7115a81d516