Isolate need for loop in player
While reading through the player code, there was a loop in the
add text track function that I found challenging to read. After
a while I found that it was just to find the period that
comes after a given period. I rewrote that section of the code
to better communicate that.

Change-Id: I659120fe309f1f5bea9152567208884a06082889