Define Active Streams Map As Its Own Object
While reading through the player code I found it hard to understand
what |activeStreamsByPeriod_| meant. After a while I found that it
was the history of which streams were last played in each period.

Knowing this I wrapped the code it a |ActiveStreamsMap| class so that
there was a documented interface for what we were doing rather than
just using a map of maps.

Change-Id: I42a6b89a2aeebd7ae3c76ad11db966bdab98f67f