Move Errors Closer To Logic
In |filterNewPeriod_| in Player, there are two possible errors that
could be thrown. However throwing the error was pretty far from where
the logic was determined.

After a while of trying to figure it out, I could not see why we
would want the error to be thrown so far away from when/where we
determined an error should be thrown.

This change moves the throwing of the error to be closer to the logic
that determined the need to throw the error.

Change-Id: If748bd7cb1b68a0d244999da3182aa8f0fdcc57a