Better errors for no persistent license support.
Beforehand, the demo would have an error saying 'This browser does
not support persistent licenses.' if the drm systems in the asset
were not supported. This error makes sense on a platform with no
persistent license support at all, but is misleading on a platform
which has persistent license support but not for the specific drm
system in the asset.
This adds separate error messages for those two cases.

Change-Id: Ib7b0eb9f7fc635867344b00fe3951abde59f538b