Notify segments on HLS manifests
Beforehand, we would call presentationTimeline.notifyMaxSegmentDuration,
which effectively does half of what presentationTimeline.notifySegments
does. At the time, we only thought that that was only what we had to do.
However, it did mean that we were never setting the
maxFirstSegmentStartTime_, which was causing HLS livestreams with
overridden segment availability times to have incorrect seek ranges.

Closes #1307

Change-Id: I3a9385eb522579156e088f298e404063bf9542dd