Pass Variants To DrmEngine
Before, we would need to pass a manifest to DrmEngine to initialize
it. This was a problem as we look to support playback via src=.

When looking at how DrmEngine used the manifest, it appears to just be
a simple way to pass in the variants and the offline session ids. By
changing DrmEngine to accept the variants and offline sessions ids, we
can break the association between DrmEngine and Manifest.

Issue #816

Change-Id: Ie0dc82aaec871d7e6e002fe59b3a984449eda8c7