Share Networking Engine From Storage
Allow users to get access to the networking engine that storage
instances are using.

This is part of allowing storage to be used without a player instance
as it will allow users to configure the networking engine when they
create a storage instance without a player instance.

The code still assumes that Storage should never destroy the networking
engine. This is something that will need to change when storage can
be initialized without player.

Issue #1297

Change-Id: If8d1642259e0cf24cb43f2ca7936227e8d939260