Split DRM Init Function
There are two intents when we initialize Drm Engine and the meaning
of the parameters are nuances depending on what you are trying
to do.

This change creates two different initializing methods so that
initializing drm engine for the intended purpose can be more
clear.

With this, some internal naming was change to better reflect how
some variables were being used to give them a bit more context.

Change-Id: Id8925eca53a34fe870ba5c973f9f0bb584f6e2d6