Support Injecting Delays Networking Engine
Our storage progress tests on Edge were sometimes failing as the promise
resolution order would change. This happened because we queue streams to be
downloaded in parallel.

This changes FakeNetworkingEngine to allow us to register an async function
as a callback for a request. This allowed us to update the test to resolve
network requests in lock step ensuring a deterministic order.

Since the progress test rely on a lot of assumptions, the tests were moved
into their own top-level group and now have their own test manifests.

Change-Id: Id366efccd15d90814ee34ca9a17df7d733ab6d67