Don't error when skipping mp4 boxes with bad size
In some situations, the mp4 parser might be required to read a box
that has a "size" value that is longer than its actual size.
For example, if there was a partial range request that ends in the
middle of a box.
Obviously, trying to READ a box like that would be a bad idea, but
if the box is going to be skipped anyway, skipping it without an
error is probably not harmful.

Closes #1535

Change-Id: I31d4c00426e6127ef6e1492e9edc7e10ac50ec8b