Isolate Drm Engine Creation In Player
In preparation to isolate the manifest filtering in player,
this change looks to isolate the creation of Drm Engine in
player.

This did require updating a couple tests as the implementation
of the test no longer reflected the implementation of the
library function. The tests were updated (and renamed) in an
attempt to better reflect the original intent.

Change-Id: I3c44ed2f3cf6c310e0f380051cfba3ae78790d87