Fix filtering of HLS audio-only content
When a piece of content has audio-only variants and audio-video
variants, we remove the audio-only ones.  We do this because we can't
drop SourceBuffers in MediaSource and switch between audio-video and
audio-only.

Some HLS content contains multiplexed streams, in which audio content
and video content are in the same segments.  In this case, our
filtering logic failed to understand that the multiplexed streams
contained both audio and video content, leading to the audio-only
content being left in, and even preferred over audio-video based on
bitrate.

This fixes the filtering logic to account for multiplexed streams, and
fixes an incidental bug in HLS parsing of multiplexed codec strings.

Change-Id: I513faa79842a506de6e969af6199de41235b4ced