Fix stalls on Tizen TV.
Sometimes the Tizen TV will stall inside a buffered range.  We have
stall-detection logic, but for Tizen's case, it doesn't change the
readyState of the video, so our logic doesn't trigger.  This changes
the logic to work for other kinds of stall too.

This also changes the tests since they waited too long.  Since the
tests would wait for a second, the stall-detection logic would trigger
and move the playhead.  So now the tests move time more frequently to
more accurately represent the moving playhead time.

Fixes: 110093558
Change-Id: If4dd36dd3da309d5a9bc67d539bf2f9cb27df862