Disable uncompiled mode on browsers without async
Browsers new enough to have ES6 arrow functions, but too old to have
async/await, should not be allowed to use uncompiled mode in the demo
app.

This also fixes a bug in which the disabled "uncompiled" link was
still clickable.

This came up while poking at Firefox 48.  To be clear, this is NOT
a supported browser version anyway, due to apparent MediaSource bugs.

Change-Id: Ic055357e5149236722afae46d0652e51edaf90d0