Always create media source for external text.
We used to only create media source for external text
when captions are turned on. It created problems with combining
external text with native controls. Native captions button changes
the state of a text track (unlike our button, which also triggers
media source creation if needed), but because there is no media source for
text, we had nothing to display.

Fixes #1596.

Backported to v2.4.x

Change-Id: Ifeca9bf987f78061d77ccbc89bf94066061ebd74