Use Tracks API to get Language And Roles
We have the Tracks API which allows app developers to get useful
information about the content we are playing. However we don't
even use it when possible.

This changes the get*Languages and get*LanguageAndRoles methods
to use the tracks API. In doing so, this actually makes each
implementation simpler.

Change-Id: I9e66cc4d4d53379b4037323f370f76f8e3d97a19