Fix MediaSource tear down race.
First, we cannot destroy the EventListener in MediaSourceEngine until
the in-progress tasks are done so we can get the 'updateend' event.
Second, we shouldn't detach MediaSource until the tasks are done to
ensure the media stack is still working.

As part of that, we need to delay the destruction of DrmEngine to
ensure it happens after MediaSourceEngine.

Fixes: 118148113
Change-Id: I2c77d72556eb0b1b6ef4a3f7a092f182b94e9156