Resolve TODO in Storage
There was a TODO I left in storage about using sets when finally
adopted them, so I am happy to say I got to resolve it to day.

The change here is that instead of using a object as a map to track
which ids we have seen, we can now use sets.

Change-Id: I9288c3e48c67d2f8cfeda70883ffb683bf735cae