Fix Image Links in "Talking About Languages"
In the "Talking About Languages" doc, the links to the images
were broken. It appears that there were minor typos in the paths
that caused images to not be found. The change updates the links
rather than the file names to avoid breaking other documents that
may reference these same images.

Issue #1691

Change-Id: I2874c8e920b1dd66f8d3a276f19953f17a999c5b