Fix chopped playback in MS Edge
Fudge the append window end to fix a playback bug in Edge.

Because Edge does not seem to implement MP4 edit lists correctly, we
may end up with some audio truncation and a gap.  If we fudge the end
of the append window a bit, we avoid this truncation and we get a
continuous buffered range and smooth playback.

Fixes #1597
Fixed #1435

Backported to v2.4.x

Change-Id: If28f3f5499dbf5ad9d33ee4859f95ebd91ad3e59