Define Deprecation Tools
Define a set of tools that we can use to track, test, and communicate
features that are being deprecated.

When you want to communicate that a feature is now deprecated, you just
need to put:
  ```
  shaka.Deprecate.enforce(MAJOR_VERSION_NUMBER,
                          MINOR_VERSION_NUMBER,
                          'Your message about what is being deprecated');
  ```

This will now print warnings when the feature is waiting to be removed and
print errors and fail asserts when the expiration has passed.

Change-Id: Ic66c53ecdfba7922cf961c7f588ab06beceb16ac