Limit scope of Edge work-around in DrmEngine
Before we had a method that would return all the supported types. The
only public use of this method was in the tests. Within DrmEngine we
would only use it to see if we could support a content type.

So instead of needing a caller to know about working around edge,
this changes it work around to be limited to one function.

This then allowed a consolidation of the other "support" methods in
DrmEngine.

Change-Id: Ia2a3484b0f8c9be51fe58955ac1ed146d4e7ba0a