Regression test for language/MediaSource failure
This is fairly tricky to reproduce.

First, we must trigger the auto-display of subtitles.  We didn't have
any tests that covered this before, so we needed this anyway.  This is
triggered by careful choice of content language and language
preferences, which required us to add language information to our
simulated test content.  I also added tests for the cases where we
should *not* trigger auto-display of subtitles.

Second, the setup phase for text must complete more quickly than the
setup for audio & video.  In real life, this happens when text is
non-segmented VTT and audio & video use DASH's SegmentBase.  In the
test, this is accomplished by delaying createSegmentIndex() of
audio and video in our simulated content.

This will cause auto-display of subtitles to trigger a setup race in
StreamingEngine.  When text wins the race, MediaSource errors follow
for audio and video.

Issue #1696

Change-Id: I1c22089925486da642368bec269a55d8556900d1