Overhaul range input elements in the UI
For the volume bar and seek bar, we use an input element of type
"range".

These elements are composed of two main parts: a "track", which is the
horizontal bar, and the "thumb", which is the knob which slides along
that bar.

In order to style the track across browsers (cough, IE 11), we need to
do something a bit tricky.  Styling the track is a nightmare,
especially if you want the thumb to be larger.  On IE 11, this gets
clipped at the track size.  So a tiny track with a large thumb is not
easily achieved.  It can be done, but the techniques for it are
incompatible with the gradient background we want to apply to it.

The solution is to put the input inside a div container, and apply the
background gradient styles to the container.  The container will act
as a visible, virtual track, inside which is contained a larger,
invisible track, in which is contained a visible thumb.  This way, the
thumb is not larger than the actual track (for IE 11's sake), but can
be larger than the virtual track.  And since we are still using a
semantically correct input element, the element is inherently
accessible.

Fixes #1703

Change-Id: I80e1c18d56e0d1455efd26310b699561bf2f41b0