Fix support for fullscreen events
When we enter or leave fullscreen without the user clicking our
fullscreen button, we should still react and update the UI as if the
button had been the cause.

This change splits up changing fullscreen state on click from reacting
to a state change, which may or may not be caused by a click.

Fixes #1725

Change-Id: I37b98c6e9eeb508a07c91fec40627e8073295557