Use Timer In Playhead Observer
Before, we would schedule (and reschedule) the poll function in the
playhead observer. We do this a lot in other classes and even have a
timer class to do that.

This change updates Playhead Observer to use the Timer class to handle
the polling events.

Change-Id: I5982ab08de47d4f8a71ab467989f63c06d0853d2
(cherry picked from commit b8a807b2ef0b2906ee3856a9b2a5dc4922973679)