Refactor StreamingEngine to not use Playhead directly
In streaming engine, we would use playhead to get the current
presentation time. Playhead was provided via the player
interface (defined by streaming engine).

Since streaming engine only needed to know what the presentation
time is, this changes the player interface to require a function
to get the presentation time rather than the playhead.

The general theme of this move is to help define the interface
in terms of needs and not the means in which they are fulfilled.

Change-Id: I442d54f08b650f22b9273a191d5a3a57af37b5bf