Use Period To Track Active Streams
Use period references rather than ids to index streams in the
active streams map.

Before, we would use the period index to index into the active
streams map, this was because we could not use object references
with Object-maps. Now, with Map, we can use references as keys.

By using the period, we don't need to get the index and the period
whenever we are interacting with the active stream map. This allowed
us to create |getPresentationPeriod_| which will give us the period
that we are currently presenting to the viewer. "Presentation" was
used to help differentiate from the period that we are currently
streaming (as they can differ).

Change-Id: I71260601271fb3ad3808b0ba86d559f88cda1e4f