Store Streams in Active Stream Map
Before we were storing the stream id and not the stream in
the active stream map. I don't know why we did this, but
storing the streams aligns better with using periods as
keys.

The only times this does not play nicely is with some of
our stream utils which ask for ids rather than streams.

Change-Id: Ic28256a2852fa4b02ae4f28f0958d393b1e2694d