Also use large gap threshold on Chromecast.
It appears that Chromecasts stop a fraction of a second before a gap.
This would sometimes cause them to stall on gaps in a way that the
stall prevention would not fix.
This changes Chromecasts to use the large gap threshold, to fix those
cases.

Issue #1720

Backported to v2.4.x

Change-Id: Ifc6720e55ab35b81b8592dad2ba7e7293983a755