Be more explicit in DrmEngine destroy
We do a lot of tear down in DrmEngine at the same time. While this often
works well for common tasks, when tasks focus on different parts, it can
introduce some race conditions.

This change looks at being a little more strict about the order we do things
in DrmEngine destroy.

While this does not fix #1728, the hope is that this will correct the
"read property 'catch' of undefined".

Issue #1728

Backported to v2.4.x

Change-Id: Id3cd25b416969ecd75a8cdd6d36959e34673fa1b