Use getSelectable* in get*Tracks
A little while ago we added the methods getSelectableVariants and
getSelectableText to create a way to get the variants and text that
could be selected. They are used with |selectAudioLanguage| and
|selectTextLanguage|.

Since we were using them for selecting variants and text, it would be
a good idea to use them when creating variant and text tracks.

A natural fallout of this change, was that the |getVariantTracks| and
|getTextTracks| methods in |StreamUtils| started to look less-and-less
important. Part of this was due to |ActiveStreamMap| storing actual
streams and not ids. So in this change, you will see that the utils have
gone away. This did require a change in |Storage| but splitting this work
across multiple CLs felt unnecessary.

Change-Id: I2d9b5b0d7b70a6b2c2e7fa51d111030cb43aa68c