Create Jasmine Matchers for OnceMore Checks
In our tests we regularly have checks that repeatedly check if a
function/spy has been called once more. To check this we would check
the call and then reset the calls each time.

To abstract this out, this CL defines two custom jasmine matchers that
allow us to check if a function/spy has been called once more.

To validate the matchers, TextEngine tests were update to make use of
it.

Change-Id: Ia6bf8d0a585648126881e1713b35be674bd47ba7