Improve language name fallback
First try to resolve the full locale, including region.
If that fails, try the base language only.
Finally, report "unknown".

When there is a loss of specificity (either to a base language or to
"unknown"), we should append the original locale in full.  Otherwise,
there may be multiple identical-looking items in the list.

Fixes #1747

Change-Id: Ie711868da8ac73cefb849158ff77d5e7c23d0927