Model Stall Detection On Its Own
While looking into a bug that involved stalling content, I found it
hard to keep the separation between stalls and gaps straight in the
code.

To help make it easier, I tried moving the stall code into its own
class (but used by the previously owning class).

During this change, an edge case was found, when paused for a while,
when we start playing again we see that as a stall. To avoid this,
we reset our stall duration so that once we start playing again, it
won't think it was stalled.

Change-Id: Ia6055e56e12ae4bc02063eeb6bbf6ffe472c2cb5