Fix percentage-width sizing for play button
This takes a completely different approach to play button sizing.  The
intended effect is the same, but this method uses padding to achieve a
1:1 aspect ratio instead of the size of the spinner contained within.

Because the spinner was sometimes hidden with display:none, it would
not take up any space in that state, causing the play button size to
collapse as well.  By using padding and overlay mixins instead, the
play button size is independent of the spinner size.

Closes #1749

Change-Id: I2e46f067283001ca2f2db44c0c6ea675ea781f46