Disable offline storage on some assets.
Not all of our assets support offline storage. For example, livestreams
inherently cannot be stored. In addition, some DRM assets do not have
the right policy settings to be stored offline.
This change marks some assets as being offline-compatible, and disables
the button if the asset is not.
It also removes some obsolete code that somehow ended up getting into
the asset structure.

Issue #1768

Backported to v2.4.x

Change-Id: I22671bd82e9c5e4645158a6e0a2f8f91893082c7