Fix video reloading on audio language change.
Previously, on seek, if any stream was unbuffered, it was considered an
unbuffered seek and every stream was cleared. This was to account for
the possibility that the streams would end up in different periods, and
therefore cause a hang.
However, this meant that changing the audio stream caused the video
stream to be flushed and reloaded pointlessly.
This changes the code to only clear all buffers if a stream has moved
into a different period; if one hasn't, it clears them on a case-by-case
basis.

Closes #1714

Change-Id: I6385e24266834bfc0c6dd2678facc5559affb941