Refactor Playhead Observers Into Separate Classes For Each Observation
Before we had one central playhead observer that would monitor multiple
different relationships:
  1. Buffering
  2. Periods
  3. Timeline Regions

This CL breaks each observer into their own class (allowing for them
to be tested independently). To allow the player to be ignorant of which
ones are in-use, a playhead observer manager has been created.

The player is responsible for creating the observers since it has the
knowledge of what ones need to be created (and the references to
everything).

As part of breaking the observers apart, the Region Timeline was broken
out so that the observer has something to observe (similar to how the
buffering observer observes the buffered region and the period observer
observes the periods).

Change-Id: I18bf0f45561f3bac44da50c0d00a02f8b1fbf3cf