Refactor player to not need manifest when selecting streams from period
When streaming engine calls back into player to pick streams we check
if the deferred selected streams are from the current period.

Before we did a round-about way to see if the streams belong to the
current period. However, we can do this in a more direct way. A
side-effect of this is we no longer need to pass the manifest into
the method.

Change-Id: I2823c36fda1e8000ae0d7f058a8cccc4a3df8c66