Refactor DownloadManager to limit internal complexity
Download manager would previously queue all the requests and
then start them after |download| was called.

By making download manager start downloading each group once
the group is created, it allowed for the internals of download
manager to be simplified.

With this simplification, it is far easier to understand how the
destroy flow works.

Change-Id: Ie2d7776eae4224e2ff1c607353c22750daa4c14f