UI: Hide control panel when mouse leaves it
Hide the control panel when the mouse leaves the control panel and the
video container.
'mouseleave' is fired when the pointer has exited the element and all of
its descendants, whereas 'mouseout' is fired when the pointer leaves the
element or leaves one of the element's descendants(even if the pointer
is still within the element).
We should hide the control panel when the mouse leaves all the elements
of the control panel, or the video container.

Closes #1687

Change-Id: I7d4c8628fd75ee0737a4a11288828524aa11089d