Refactor UI Controls to Use Timer
We have a timer class in shaka.util.Timer to make it easier to
work with repeating time-based events.

This CL changes shaka.ui.Controls to make use of shaka.util.Timer.
To ensure that delayed vs immediate invocations of the events did
not conflict with each other, a new method (tick) was added to
shaka.util.Timer.

|tick| works like |start|, but in an immediate synchronous form.

Change-Id: I50f5ac35d72f58bc55fcc9d20811dbc3dd5ba55c